# backend/routes/api-docs/swaggerOptions.json
{
	"openapi": "3.0.2",
	"info": {
		"title": "Tourism Trends API",
		"description": "The API endpoints and their usage are described in detail below.",
		"version": "1.0.0",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "glenjplayer@gmail.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tags": [
		{
			"name": "Trends",
			"description": "API queries based tourism trend data."
		},
		{
			"name": "User Data",
			"description": "API queries based on user data."
		},
		{
			"name": "Authentication",
			"description": "User login and registration."
		}
	],
	"paths": {
		"/trends": {
			"post": {
				"tags": ["Trends"],
				"description": "Returns a list of trend data. The list is arranged by date, in ascending order.",
				"operationId": "listAllMatchingTrends",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "region",
						"in": "body",
						"required": false,
						"description": "Local Government Area by id in array. Defaults to all regions if not provided for Local Government Memebers, defaults to the region of the user for Business Members.",
						"schema": {
							"type": "array"
						}
					},
					{
						"name": "type",
						"in": "body",
						"required": false,
						"description": "Type of trend data in array. Options are ['ave_historical_occupancy', 'ave_daily_rate', 'ave_length_stay', 'ave_booking_window']. Defaults to all types if not provided.",
						"schema": {
							"type": "array"
						}
					},
					{
						"name": "dateRange",
						"in": "body",
						"required": false,
						"description": "Date range in the format of an array ['yyyy-mm-dd', 'yyyy-mm-dd']. Defaults to last two years.",
						"schema": {
							"type": "array"
						}
					}
				],
				"responses": {
					"200": {
						"description": "An array of objects containing id, date, lga_id, lga_name, ave_historical_occupany, ave_daily_rate, ave_length_stay, ave_booking_window. An example of one object in the array is shown below.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										},
										"query": {
											"type": "object",
											"properties": {
												"region": {
													"type": "array"
												},
												"type": {
													"type": "array"
												},
												"dateRange": {
													"type": "array"
												}
											}
										},
										"results": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": {
														"type": "number"
													},
													"date": {
														"type": "string"
													},
													"lga_id": {
														"type": "number"
													},
													"lga_name": {
														"type": "string"
													},
													"ave_historical_occupancy": {
														"type": "number"
													},
													"ave_daily_rate": {
														"type": "number"
													},
													"ave_length_stay": {
														"type": "number"
													},
													"ave_booking_window": {
														"type": "number"
													}
												}
											}
										}
									}
								},
								"example": {
									"error": false,
									"message": "Success",
									"query": {
										"region": [1, 2],
										"type": ["ave_historical_occupancy"],
										"dateRange": ["2024-01-01", "2024-01-31"]
									},
									"results": [
										{
											"id": 1461,
											"date": "2024-01-01",
											"lga_id": 1,
											"lga_name": "Cairns",
											"ave_historical_occupancy": 0.7554
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid type, region or dateRange provided.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "No region provided."
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Internal server error"
								}
							}
						}
					}
				}
			}
		},
		"/user/trends/get": {
			"post": {
				"tags": ["User Data"],
				"description": "Use to request user data, previously imported.",
				"parameters": [
					{
						"name": "type",
						"in": "body",
						"required": false,
						"description": "Type of trend data in array. Options are ['ave_historical_occupancy', 'ave_daily_rate', 'ave_length_stay', 'ave_booking_window']. Defaults to all types if not provided.",
						"schema": {
							"type": "array"
						}
					},
					{
						"name": "dateRange",
						"in": "body",
						"required": false,
						"description": "Date range in the format of an array ['yyyy-mm-dd', 'yyyy-mm-dd']. Defaults to last two years.",
						"schema": {
							"type": "array"
						}
					}
				],
				"responses": {
					"200": {
						"description": "An array of objects containing id, date, lga_id, lga_name, ave_historical_occupany, ave_daily_rate, ave_length_stay, ave_booking_window. An example of one object in the array is shown below.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										},
										"query": {
											"type": "object",
											"properties": {
												"region": {
													"type": "array"
												},
												"type": {
													"type": "array"
												},
												"dateRange": {
													"type": "array"
												}
											}
										},
										"results": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": {
														"type": "number"
													},
													"date": {
														"type": "string"
													},
													"lga_id": {
														"type": "number"
													},
													"lga_name": {
														"type": "string"
													},
													"ave_historical_occupancy": {
														"type": "number"
													},
													"ave_daily_rate": {
														"type": "number"
													},
													"ave_length_stay": {
														"type": "number"
													},
													"ave_booking_window": {
														"type": "number"
													}
												}
											}
										}
									}
								},
								"example": {
									"error": false,
									"message": "Success",
									"query": {
										"region": [1, 2],
										"type": ["ave_historical_occupancy"],
										"dateRange": ["2024-01-01", "2024-01-31"]
									},
									"results": [
										{
											"id": 1461,
											"date": "2024-01-01",
											"lga_id": 1,
											"lga_name": "Cairns",
											"ave_historical_occupancy": 0.7554
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid type, region or dateRange provided.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "No region provided."
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Internal server error"
								}
							}
						}
					}
				}
			}
		},
		"/user/trends/add": {
			"post": {
				"tags": ["User Data"],
				"description": "Use to add user specific data for comparison to public data.",
				"parameters": [
					{
						"name": "data",
						"in": "body",
						"required": true,
						"schema": {
							"type": "array"
						}
					}
				],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"data": {
										"type": "array",
										"items": {
											"type": "object",
											"properties": {
												"date": {
													"type": "string"
												},
												"lga_id": {
													"type": "number"
												},
												"ave_historical_occupancy": {
													"type": "number"
												},
												"ave_daily_rate": {
													"type": "number"
												},
												"ave_length_stay": {
													"type": "number"
												},
												"ave_booking_window": {
													"type": "number"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ".",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": false,
									"message": "Success"
								}
							}
						}
					},
					"400": {
						"description": "Invalid query parameters.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Error."
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Internal Server Error"
								}
							}
						}
					}
				}
			}
		},
		"/user/register": {
			"post": {
				"tags": ["Authentication"],
				"description": "Use to register a new user.",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								}
							},
							"example": {
								"email": "test@test.com",
								"password": "password"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "The user has been successfully registered.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"message": "User created"
								}
							}
						}
					},
					"400": {
						"description": "Bad request.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Request body incomplete, both email and password are required"
								}
							}
						}
					},
					"409": {
						"description": "Conflict. User already exists.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "User already exists"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Internal server error"
								}
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"tags": ["Authentication"],
				"description": "Use to login a user.",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								}
							},
							"example": {
								"email": "test@test.com",
								"password": "password"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "The user has been successfully logged in.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"token": {
											"type": "string"
										},
										"token_type": {
											"type": "string"
										},
										"expires_in": {
											"type": "number"
										}
									}
								},
								"example": {
									"token": "aJWTBearerToken",
									"token_type": "Bearer",
									"expires_in": 86400
								}
							}
						}
					},
					"400": {
						"description": "Bad request.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Request body incomplete, both email and password are required"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized. Invalid email or password.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										},
										"message": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Incorrect email or password"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									}
								},
								"example": {
									"error": true,
									"message": "Internal server error"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter JWT Bearer token only"
			}
		}
	}
}